refactor: extract fetchOfferHits to a new file
