+++ src/libs/algolia/fetchAlgolia/fetchOffer.ts
@@ -1,6 +1,5 @@
 import { Hit } from '@algolia/client-search'
 
-import { initialSearchState } from 'features/search/pages/reducer'
 import { Response } from 'features/search/pages/useSearchResults'
 import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
 import { client } from 'libs/algolia/fetchAlgolia/clients'
@@ -33,32 +32,5 @@ export const fetchOffer = async (
   } catch (error) {
     captureAlgoliaError(error)
     return { hits: [] as Hit<SearchHit>[], nbHits: 0, page: 0, nbPages: 0 }
-  }
-}
-
-export const fetchOfferHits = async (
-  objectIds: string[],
-  isUserUnderage: boolean
-): Promise<SearchHit[]> => {
-  const index = client.initIndex(env.ALGOLIA_OFFERS_INDEX_NAME)
-  const searchParameters = buildOfferSearchParameters(
-    { ...initialSearchState, hitsPerPage: objectIds.length, objectIds, query: '' },
-    null,
-    isUserUnderage
-  )
-
-  try {
-    const response = await index.search<SearchHit>('', {
-      page: 0,
-      hitsPerPage: objectIds.length,
-      ...searchParameters,
-      attributesToRetrieve: offerAttributesToRetrieve,
-      attributesToHighlight: [], // We disable highlighting because we don't need it
-    })
-    const hits = response.hits.filter(Boolean) as SearchHit[]
-    return hits.filter(({ offer }) => !offer.isEducational)
-  } catch (error) {
-    captureAlgoliaError(error)
-    return [] as SearchHit[]
   }
 }
