+++ src/libs/algolia/fetchAlgolia/fetchOfferHits.ts
@@ -1,0 +1,36 @@
+import { initialSearchState } from 'features/search/pages/reducer'
+import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
+import { client } from 'libs/algolia/fetchAlgolia/clients'
+import {
+  offerAttributesToRetrieve,
+  buildOfferSearchParameters,
+} from 'libs/algolia/fetchAlgolia/config'
+import { env } from 'libs/environment'
+import { SearchHit } from 'libs/search'
+
+export const fetchOfferHits = async (
+  objectIds: string[],
+  isUserUnderage: boolean
+): Promise<SearchHit[]> => {
+  const index = client.initIndex(env.ALGOLIA_OFFERS_INDEX_NAME)
+  const searchParameters = buildOfferSearchParameters(
+    { ...initialSearchState, hitsPerPage: objectIds.length, objectIds, query: '' },
+    null,
+    isUserUnderage
+  )
+
+  try {
+    const response = await index.search<SearchHit>('', {
+      page: 0,
+      hitsPerPage: objectIds.length,
+      ...searchParameters,
+      attributesToRetrieve: offerAttributesToRetrieve,
+      attributesToHighlight: [], // We disable highlighting because we don't need it
+    })
+    const hits = response.hits.filter(Boolean) as SearchHit[]
+    return hits.filter(({ offer }) => !offer.isEducational)
+  } catch (error) {
+    captureAlgoliaError(error)
+    return [] as SearchHit[]
+  }
+}
